HotFix and New Features
Some bugs was fixed, the next update I will insert the try/exception, but on this the try/except is difficulty my work haha.
I'm work on a way to retrieve the data to avoid we to make a API request to azure resource. It is basically re-build the data that is on the database. so its a little complicated.