Introducing entities_inputs tables
The biggest challenge so far.
I had to change the way that we insert the None subcategory into database, now it is NULL and other big change is the solve of the problem that imagine that you have this sentence, "My dad's name is ball, I Have a ball" in this phrase ball is two different entities, a Person and a Product , the databases before this commit just insert the first entities type, now we can have more entities with the same name, but different categories and sub-categories.